Temporarily remove tests that depend on tld ignoring.
Ignoring valid top level domains to avoid false positives is probably not viable,
since .zip and .mov are top level domains that are likely to be both false positives
and used maliciously. Another method will have to be used to avoid script syntax
that happens to collide with a top level domain.