Minimum size for a domain found on its own
Currently the minimum domain length accepted is 4,
c.td where c is some character and td is a country code.
This is too short and can appear randomly in data.
Adding a minimum of 6 (ccc.td) since 1 and 2 character second level domains
are more limited (used to be reserved). This only applies to finding a domain raw
and not as part of a url